Cleanup and bugfixes in the new database layer, elsewhere in the framework relating to the database and wishlist example application. The example is now again fully functional.
What used to be one() in DAOs are now removed, as load() now functions as both,
that is it sets the model into the proxy as well as returning the plain model
object.

A new SqlException class is introduced, currently only thrown in query() in
database connection implementations and never catched. It made for easier
debugging of bugs, and will be cleaned up and better integrated in coming
commits.