DateTimeValidator.php: Fixed bug which made mktime() barf, by casting parameters to int.
EmailValidator.php:
Fixed bug which prevented the correct property from being validated.

HexValidator.php:
Fixed bug which caused invalid hex numbers to validate.

UniqueValidator.php:
Added support for a 'dbType' validation parameter, to enforce the correct type
of the column in the database when doing the validation. This was made to
prevent cases where a text column could contain numbers (i.e. phone numbers)
which would incorrectly be treated as an integer in the query.
Also added support for NULL values in the WHERE clause, as NULL values must
be compared using IS/IS NOT instead of =/<>.