Completed support for defining interceptor settings in application config files, which overrides default settings in core interceptor configuration.
Logging of errors and exceptions also work again, which also respects the
current environment mode now. Details of an error or exception will never
be displayed to the client in production mode. But logging to file and/or
email can be turned on/off independently for any environment.

A lot of loading and preparation of configuration values have been extracted
to a separate ConfigHelper class. The main Config class is thus maintained
mostly as a container and static interface. The groundwork have also been
laid for caching the fully loaded Config instance.

Sample production.ini configuration file is now provided, similar to the old
config.php that was included. Placeholder development.ini and test.ini with
a short introductionary comment also exists.

Updating the example wishlist application remains to be done.