Rewritten DefaultActionMapper to support "nested routes"; URIs with IDs relating to specific "sections" of the URI logically positioned in the URI. For instance, you can now have an URI "/lists/1/contacts" which would map to the contacts index action for the list with ID 1, which in this example is set into the request parameter "listsid".
A new default action "View" alongside the existing "Index" is also now
supported, which will be used if a directory was the last match, and an ID has
been found in the URI. Following the same example, with an URI
"/lists/1/contacts/4" would call the contacts view action, where the ID of the
contact is in the parameter "contactsid".

Naming of action classes has also been changed. Instead of only naming them
based on the last part of the URI, the complete "path" must now be in the class
name. Instead of /lists/contacts/Index.php, it must now be named
/lists/contacts/ListsContactsIndex.php. This was done to fascilitate testing of
multiple actions, where class names must not collide.