Add phone validator
The phone validator uses the PHP port of Google's `libphonenumber`. It
supports defining the region for validating numbers within and optional
type to attempt validating. Region is not required if phone numbers
contain the country code, but must otherwise be provided. Type can not
be guaranteed verified. Currently only `mobile` and `landline` are valid
types to specify.