Automatically reconnect PostgreSQL database connections
Before issuing a query we always make sure a database connection is set
up. But if this connection has become broken the query will fail
anyhow. This performs a ping on any existing connection before queries
are attempted, which will explicitly check the connection and
automatically try to reconnect if the connection is broken. If the ping
with attemped reconnect fails, we try to connect fresh again too. If
even that fails, there's something seriously wrong which we can't
automatically recover from and the following query will simply throw an
exception as before.