Support encoding JSON columns in PostgreSQL
The `excapeValue` function for PostgreSQL is updated, to further inspect
array values. If the array is a plain sequential array we encode it as a
regular PostgreSQL array as before, whereas if the array is an
associated array (i.e. with non-integer keys) we assume it's supposed to
be represented in a JSON(B) column and therefore JSON encode the value.