Improve `PhoneValidator` with strict type check and formatting
The use of `isPossibleNumberForTypeWithReason` didn't properly check the
type of the phone number. It checked that the length of number matched
the type. So in our case that would always be true for landline numbers
and mobile numbers. To fix this we make sure the type is properly
checked and we also support formatting the phone number if specified.